+++ frontend/src/Dashboard/AdminDashboard.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { motion } from 'framer-motion';
 import { Search, MoreVertical } from 'lucide-react';
 import HeaderRightSection from '../components/HeaderRightSection.jsx';
-import Layout from './Layout.jsx';
+import Layout from '../components/Layout.jsx';
 
 
 const Dashboard = () => {
