+++ frontend/src/layout/EventManagement.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Layout from "./Layout.jsx";
+import Layout from "../components/Layout.jsx";
 
 const EventManagement = () => {
   const navigate = useNavigate();
