+++ frontend/src/layout/TeacherDetail.jsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { ArrowLeft, MessageCircle, Edit2, X, Save, Send } from 'lucide-react';
 import { motion } from 'framer-motion';
-import Layout from './Layout.jsx';
+import Layout from '../components/Layout.jsx';
 
 const TeacherDetail = () => {
   const navigate = useNavigate();
